XWiki-9151 : XSS vulnerability in link syntax when using "path:". Creating a SecureXMLWikiPrinter
