+++ xwiki-rendering-api/src/main/java/org/xwiki/rendering/renderer/printer/XHTMLWikiPrinter.java
@@ -19,11 +19,6 @@
  */
 package org.xwiki.rendering.renderer.printer;
 
-import java.util.Map;
-
-import org.apache.commons.lang3.StringUtils;
-import org.xml.sax.Attributes;
-
 /**
  * Base toolkit class for all XHTML-based renderers. This printer handles whitespaces so that it prints "&nbsp;" when
  * needed (i.e. when the spaces are at the beginning or at the end of an element's content or when there are more than 1
@@ -32,177 +27,7 @@ import org.xml.sax.Attributes;
  * @version $Id$
  * @since 1.7M1
  */
-public class XHTMLWikiPrinter extends XMLWikiPrinter
+public interface XHTMLWikiPrinter extends XMLWikiPrinter
 {
-    private int spaceCount;
-
-    private boolean isInCData;
-
-    private boolean isInPreserveElement;
-
-    private boolean elementEnded;
-
-    private boolean hasTextBeenPrinted;
-
-    /**
-     * @param printer the object to which to write the XHTML output to
-     */
-    public XHTMLWikiPrinter(WikiPrinter printer)
-    {
-        super(printer);
-    }
-
-    @Override
-    public void printXML(String str)
-    {
-        handleSpaceWhenInText();
-        super.printXML(str);
-        this.hasTextBeenPrinted = true;
-    }
-
-    @Override
-    public void printXMLElement(String name)
-    {
-        handleSpaceWhenStartElement();
-        super.printXMLElement(name);
-    }
-
-    @Override
-    public void printXMLElement(String name, String[][] attributes)
-    {
-        handleSpaceWhenStartElement();
-        super.printXMLElement(name, attributes);
-    }
-
-    @Override
-    public void printXMLElement(String name, Map<String, String> attributes)
-    {
-        handleSpaceWhenStartElement();
-        super.printXMLElement(name, attributes);
-    }
-
-    @Override
-    public void printXMLStartElement(String name)
-    {
-        handleSpaceWhenStartElement();
-        super.printXMLStartElement(name);
-    }
-
-    @Override
-    public void printXMLStartElement(String name, String[][] attributes)
-    {
-        handleSpaceWhenStartElement();
-        super.printXMLStartElement(name, attributes);
-    }
-
-    @Override
-    public void printXMLStartElement(String name, Map<String, String> attributes)
-    {
-        handleSpaceWhenStartElement();
-        super.printXMLStartElement(name, attributes);
-    }
-
-    @Override
-    public void printXMLStartElement(String name, Attributes attributes)
-    {
-        handleSpaceWhenStartElement();
-        super.printXMLStartElement(name, attributes);
-    }
-
-    @Override
-    public void printXMLEndElement(String name)
-    {
-        handleSpaceWhenEndlement();
-        super.printXMLEndElement(name);
-        this.elementEnded = true;
-    }
-
-    @Override
-    public void printXMLComment(String content)
-    {
-        printXMLComment(content, false);
-    }
-
-    @Override
-    public void printXMLComment(String content, boolean escape)
-    {
-        handleSpaceWhenStartElement();
-        super.printXMLComment(content, escape);
-        this.elementEnded = true;
-    }
-
-    @Override
-    public void printXMLStartCData()
-    {
-        handleSpaceWhenStartElement();
-        super.printXMLStartCData();
-    }
-
-    @Override
-    public void printXMLEndCData()
-    {
-        handleSpaceWhenEndlement();
-        super.printXMLEndCData();
-    }
-
-    /**
-     * This method should be used to print a space rather than calling <code>printXML(" ")</code>.
-     */
-    public void printSpace()
-    {
-        this.spaceCount++;
-    }
-
-    @Override
-    public void printRaw(String raw)
-    {
-        handleSpaceWhenStartElement();
-        super.printRaw(raw);
-        this.elementEnded = true;
-    }
-
-    private void handleSpaceWhenInText()
-    {
-        if (this.elementEnded || this.hasTextBeenPrinted) {
-            handleSpaceWhenStartElement();
-        } else {
-            handleSpaceWhenEndlement();
-        }
-    }
-
-    private void handleSpaceWhenStartElement()
-    {
-        // Use case: <tag1>something <tag2>...
-        // Use case: <tag1>something <!--...
-        if (this.spaceCount > 0) {
-            if (!this.isInCData && !this.isInPreserveElement) {
-                // The first space is a normal space
-                super.printXML(" ");
-                for (int i = 0; i < this.spaceCount - 1; i++) {
-                    printEntity("&nbsp;");
-                }
-            } else {
-                super.printXML(StringUtils.repeat(' ', this.spaceCount));
-            }
-        }
-        this.spaceCount = 0;
-        this.elementEnded = false;
-        this.hasTextBeenPrinted = false;
-    }
-
-    private void handleSpaceWhenEndlement()
-    {
-        // Use case: <tag1>something </tag1>...
-        // All spaces are &nbsp; spaces since otherwise they'll be all stripped by browsers
-        if (!this.isInCData && !this.isInPreserveElement) {
-            for (int i = 0; i < this.spaceCount; i++) {
-                printEntity("&nbsp;");
-            }
-        } else {
-            super.printXML(StringUtils.repeat(' ', this.spaceCount));
-        }
-        this.spaceCount = 0;
-        this.elementEnded = false;
-        this.hasTextBeenPrinted = false;
-    }
+    public void printSpace();
 }
