+++ xwiki-rendering-syntaxes/xwiki-rendering-syntax-annotatedxhtml/src/main/java/org/xwiki/rendering/internal/renderer/xhtml/AnnotatedXHTMLChainingRenderer.java
@@ -51,7 +51,7 @@ public class AnnotatedXHTMLChainingRenderer extends XHTMLChainingRenderer
     public AnnotatedXHTMLChainingRenderer(XHTMLLinkRenderer linkRenderer,
         XHTMLImageRenderer imageRenderer, ListenerChain listenerChain)
     {
-        super(linkRenderer, imageRenderer, listenerChain);
+        super(linkRenderer, imageRenderer, listenerChain, "default");
 
         this.macroRenderer = new XHTMLMacroRenderer();
     }
